Update logging with new config management loggers
Updated the configs for each stage. Refactored the
logging config file to be a dictConfig(), much cleaner.
Add a flag for debugging seperate from normal logging so
you can turn on debug logging and it won't impact normal
logs. This is done with a simple filter attached to each
log handler and a flag in the config files. Path of
debug log is also in the config files. Yay config files!